// src/states/model/auxiliaire-justice.model.ts
export interface AuxiliaireJusticeModel {
    autAuxilRang?: number
    autAuxilDate?: string;
    auxiCode?: string;
    typauxiCode?: string;
    auxiNom?: string;
    auxiPren?: string;
    auxiAdrpost?: string;
    auxiTel?: string;
    auxiDatnaiss?: string;
    civCode?: string;
    auxiCptCode?: string;
    auxiSitgeo?: string;
}